# src/openapi/swagger.yaml
openapi: 3.0.0
info:
  title: Royal Burger API
  description: API para o sistema de delivery da hamburgueria Royal Burger.
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:5000/api
    description: Servidor de Desenvolvimento Local
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
    MessageResponse:
      type: object
      properties: {msg: {type: string, example: "Operação realizada com sucesso."}}
...
    Customer:
      type: object
      properties: {id: {type: integer, example: 3}, full_name: {type: string, example: "João da Silva"}, email: {type: string, format: email, example: "joao.silva@email.com"}, phone_number: {type: string, example: "14999998888"}, role: {type: string, example: "customer"}, is_active: {type: boolean, example: true}}
    NewCustomer:
      type: object
      required: [full_name, email, password, password_confirmation, date_of_birth]
      properties:
        full_name: {type: string, example: "Maria Oliveira"}
        email: {type: string, format: email, example: "maria.oliveira@email.com"}
        password: {type: string, example: "SenhaForte123"}
        password_confirmation: {type: string, example: "SenhaForte123"}
        phone_number: {type: string, example: "14988887777"}
        date_of_birth: {type: string, format: date, example: "1995-05-20"}
    UpdateCustomer:
      type: object
      properties:
        full_name: {type: string, example: "João da Silva Santos"}
        email: {type: string, format: email, example: "joao.santos@email.com"}
        phone: {type: string, example: "14999991234"}
        cpf: {type: string, example: "12345678900"}
        date_of_birth: {type: string, format: date, example: "1990-07-22"}
    Address:
      type: object
      properties: {id: {type: integer, example: 1}, street: {type: string, example: "Rua das Flores"}, number: {type: string, example: "123"}, complement: {type: string, example: "Apto 4B"}, neighborhood: {type: string, example: "Centro"}, city: {type: string, example: "Lins"}, state: {type: string, example: "SP"}, zip_code: {type: string, example: "16400-000"}}
    NewAddress:
      type: object
      required: [street, number, neighborhood, city]
      properties: {street: {type: string, example: "Avenida Brasil"}, number: {type: string, example: "1000"}, complement: {type: string}, neighborhood: {type: string, example: "Jardim Americano"}, city: {type: string, example: "Lins"}, state: {type: string, example: "SP"}, zip_code: {type: string, example: "16400-123"}}
    LoyaltyBalance:
      type: object
      properties: {accumulated_points: {type: integer, example: 500}, spent_points: {type: integer, example: 100}, current_balance: {type: integer, example: 400}}
    LoyaltyHistoryEntry:
      type: object
      properties: {points: {type: integer, example: 50}, reason: {type: string, example: "Pedido ID: 123"}, date: {type: string, format: date-time, example: "2025-09-15T20:30:00"}}
    Ingredient:
      type: object
      properties: {id: {type: integer, example: 25}, name: {type: string, example: "Queijo Cheddar Fatiado"}, price: {type: number, format: float, example: 1.50}, is_available: {type: boolean, example: true}}
    NewIngredient:
      type: object
      required: [name]
      properties: {name: {type: string, example: "Cebola Roxa"}, price: {type: number, format: float, example: 0.50}}
    UpdateIngredientAvailability:
      type: object
      required: [is_available]
      properties: {is_available: {type: boolean, example: false}}
...
    ProductIngredient:
      type: object
      properties:
        ingredient_id: {type: integer, example: 15}
        name: {type: string, example: "Picles"}
        quantity: {type: number, format: float, example: 4.0, description: "A quantidade do ingrediente na receita (ex: 2.0 para duas fatias)"}
    NewProductIngredient:
      type: object
      required: [ingredient_id, quantity]
      properties: {ingredient_id: {type: integer, example: 10}, quantity: {type: number, format: float, example: 2.0}}
...
security:
  - bearerAuth: []
paths:
...
  /customers:
    post:
      summary: Cria um novo cliente (cadastro).
      security: []
      tags: [Clientes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCustomer'
      responses:
        '201':
          description: Cliente criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    get:
      summary: (Admin) Lista todos os clientes.
      tags: [Clientes]
      responses:
        '200':
          description: Lista de clientes.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Customer'}
  /customers/{user_id}:
    get:
      summary: Busca dados de um cliente específico.
      tags: [Clientes]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dados do cliente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    put:
      summary: Atualiza os dados de um cliente.
      description: |-
        Atualiza um ou mais campos do perfil de um cliente. 
        A API valida o formato do CPF e a unicidade do e-mail e telefone antes de salvar.
        Um cliente só pode atualizar seus próprios dados, a menos que o solicitante seja um admin.
      tags: [Clientes]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomer'
      responses:
        '200':
          description: Dados atualizados com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: "Requisição Inválida / Erro de Validação. Ocorre se os dados enviados são inválidos (ex: CPF incorreto, e-mail ou telefone já em uso)."
        '403':
          description: Acesso proibido.
        '404':
          description: Cliente não encontrado.
    delete:
      summary: Inativa a conta de um cliente.
      tags: [Clientes]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Conta inativada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /customers/{user_id}/addresses:
    get:
      summary: Lista os endereços de um cliente.
      tags: [Endereços]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de endereços.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Address'}
    post:
      summary: Adiciona um novo endereço para um cliente.
      tags: [Endereços]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAddress'
      responses:
        '201':
          description: Endereço criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /customers/addresses/{address_id}:
    put:
      summary: Atualiza um endereço específico.
      tags: [Endereços]
      parameters:
        - name: address_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAddress'
      responses:
        '200':
          description: Endereço atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    delete:
      summary: Deleta um endereço específico.
      tags: [Endereços]
      parameters:
        - name: address_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Endereço deletado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /customers/{user_id}/loyalty/balance:
    get:
      summary: Consulta o saldo de pontos de fidelidade.
      description: Consulta o saldo de pontos do cliente. Importante - Esta rota verifica e zera o saldo se os pontos tiverem expirado por inatividade (+60 dias).
      tags: [Fidelidade]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Saldo de pontos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyBalance'
  /customers/{user_id}/loyalty/history:
    get:
      summary: Consulta o histórico de pontos de fidelidade.
      description: Retorna o histórico de transações de pontos, incluindo ganhos, resgates e expirações por inatividade.
      tags: [Fidelidade]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Histórico de transações de pontos.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/LoyaltyHistoryEntry'}
  /ingredients:
    get:
      summary: (Admin) Lista todos os ingredientes.
      tags: [Ingredientes]
      responses:
        '200':
          description: Lista de ingredientes.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Ingredient'}
    post:
      summary: (Admin) Cria um novo ingrediente.
      tags: [Ingredientes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewIngredient'
      responses:
        '201':
          description: Ingrediente criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
  /ingredients/{ingredient_id}:
    put:
      summary: (Admin) Atualiza um ingrediente.
      tags: [Ingredientes]
      parameters:
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewIngredient'
      responses:
        '200':
          description: Ingrediente atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    delete:
      summary: (Admin) Inativa um ingrediente.
      tags: [Ingredientes]
      parameters:
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingrediente inativado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /ingredients/{ingredient_id}/availability:
    patch:
      summary: (Admin) Atualiza a disponibilidade de um ingrediente.
      tags: [Ingredientes]
      parameters:
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIngredientAvailability'
      responses:
        '200':
          description: Disponibilidade atualizada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /products/{product_id}/ingredients:
    get:
      summary: Lista os ingredientes de um produto.
      security: []
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de ingredientes do produto.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ProductIngredient'}
    post:
      summary: (Admin) Adiciona um ingrediente a um produto.
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProductIngredient'
      responses:
        '201':
          description: Ingrediente associado/atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /products/{product_id}/ingredients/{ingredient_id}:
    delete:
      summary: (Admin) Remove um ingrediente de um produto.
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingrediente desassociado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...